cni: add ca-certifcates dependency
Many of the CNI plugins require authenticated connections, as such
they are looking for elements of ca-certificates.

CNI isn't small, so we add this as a general rdepends. If we need to
slim things down in the future, we can split the CNI into specific
implementations and add the dependency to those packages.

Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>