lxc: upgrade 4.0.12 -> 5.0.1
autotools => meson
Drop lxc-fix-B-S.patch
Drop lxc-doc-upgrade-to-use-docbook-3.1-DTD.patch

LXC 5.0 LTS has been released [1]
The LXC team is pleased to announce the release of LXC 5.0.0!

This is the result of two years of work since the LXC 4.0.0 release and is
the fifth LTS release for the LXC project. This release will be supported
until June 2027.

Major changes
-------------
Switch to meson
With this release of LXC, autotools is being replaced by meson as the build
tooling. Compatibility Makefile targets are provided for all, install and
dist.

This is a change which is particularly relevant for packagers as it
otherwise has no user visible impact.

LXC 5.0.1 has been released [2]
Changelog for 5.0.1:
* meson: Generate compile commands by iterating over an array
* Fix uninitialized read in parse_cap when libcap is not used
* use systemd dbus StartTransientUnit for unpriv cgroup2
* fix for issue 4026: set broadcast to 0.0.0.0 for /31 and /32
* conf: log file descriptors on error during idmapped mount setup
* start: don't overwrite file descriptors during namespace preservation
* start: record inherited namespaces earlier to make it available for
  idmapped rootfs setup
* conf: fix append_ttyname()
* start: fix namespace sharing
* add check for statvfs
* Fix off-by-one error constructing mount options
* Store mount options in correct variable
* meson: add remaining still-in-use config checks
* src/lxc/log.h: fix STRERROR_R_CHAR_P
* meson.build: fix build with -Dcapabilities=false
* meson.build: fix build without stack-protector
* README: update security mails

[1] https://linuxcontainers.org/lxc/news/#lxc-50-lts-has-been-released
[2] https://linuxcontainers.org/lxc/news/#lxc-501-has-been-released

Signed-off-by: Tim Orling <tim.orling@konsulko.com>